python-service/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
youtube-transcript-api==1.2.2
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.5.0
psycopg2-binary==2.9.10